RDKB-11866 : adding missing changes to moca_hal.h
Reason for change: adding missing changes to moca_hal.h
Test Procedure: Query Device.MoCA.Interface.1.AssociatedDevice.1.RxPowerLevel
It should give negative value also.
Risks: None

+++ moca_hal.h
@@ -367,7 +367,7 @@ typedef struct {
 	/* Tx Power Reduced by this Node */
     ULONG                           TxPowerControlReduction;
 	/* Rx Power Level read by this Node */
-    ULONG                           RxPowerLevel;
+    INT                             RxPowerLevel;
 	/* Tx Broadcast PHY Rate */
     ULONG                           TxBcastRate;
 	/* Rx Broadcast Power Level read by this Node */
